+++ ubi-nginx/test.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -e
+
+# we don't have any tests for this image, so exit
+exit 0